requirements.txt
pypdf>=4.0.0
ragas>=0.1.7
openai>=1.10.0
python-dotenv>=1.0.0
pandas>=2.1.0
tqdm>=4.66.0
jinja2>=3.1.0
chromadb>=0.4.25
sentence-transformers>=2.2.0
anthropic>=0.30.0